Updated file structure for KMP
