+++ .idea/artifacts/data.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="data">
+    <output-path>$PROJECT_DIR$/data/build/libs</output-path>
+    <root id="archive" name="data.jar" />
+  </artifact>
+</component>
