+++ .idea/artifacts/data_jvm.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="data-jvm">
+    <output-path>$PROJECT_DIR$/data/build/libs</output-path>
+    <root id="archive" name="data-jvm.jar">
+      <element id="module-output" name="Valiutchik.data.jvmMain" />
+    </root>
+  </artifact>
+</component>
