+++ data/src/androidDeviceTest/kotlin/fobo66/valiutchik/core/model/datasource/FormattingDataSourceImplTest.kt
