+++ data/src/androidHostTest/kotlin/fobo66/valiutchik/core/DataKoinModuleTest.kt
