+++ data/src/androidMain/AndroidManifest.xml
