+++ data/src/commonMain/kotlin/fobo66/valiutchik/core/entities/CurrencyRatesLoadFailedException.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2024 Andrey Mukamolov
+ *    Copyright 2025 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,18 +14,6 @@
  *    limitations under the License.
  */
 
-package fobo66.valiutchik.core
+package fobo66.valiutchik.core.entities
 
-/**
- * Constants used in project
- * Created by fobo66 on 09.08.2015.
- */
-const val BUY_COURSE = true
-const val SELL_COURSE = false
-
-const val TIMESTAMP = "fobo66.exchangecourcesbelarus.TIMESTAMP"
-
-const val KEY_DEFAULT_CITY = "default_city"
-const val KEY_UPDATE_INTERVAL = "update_interval"
-
-const val UNKNOWN_COURSE = "-"
+class CurrencyRatesLoadFailedException(throwable: Throwable) : Exception(throwable)
