+++ data/src/commonMain/kotlin/fobo66/valiutchik/core/entities/Location.kt
