+++ data/src/commonMain/kotlin/fobo66/valiutchik/core/entities/OpenSourceLicensesItem.kt
